adding redux search
