+++ src/actions.js
@@ -1,6 +1,6 @@
 import { CHANGE_SEARCH_FIELD } from './constants'
 
-export const setSearchFIeld = (text) => ({
+export const setSearchField = (text) => ({
     type: CHANGE_SEARCH_FIELD,
     payload: text
 })
